Split table specific parseutils into separate subpackage
